algorithm/leetcode/301-350/328_Odd_Even_Linked_List/328.go
package _28_Odd_Even_Linked_List

func oddEvenList(head *ListNode) *ListNode {
	if head == nil || head.Next == nil {
		return head
	}
	odd, even, evenHead := head, head.Next, head.Next
	for even != nil && even.Next != nil {
		odd.Next, even.Next = odd.Next.Next, even.Next.Next
		odd, even = odd.Next, even.Next
	}
	odd.Next = evenHead
	return head
}

func oddEvenList1(head *ListNode) *ListNode {
	if head == nil || head.Next == nil {
		return head
	}
	odd := head
	even := head.Next
	evenHead := even
	for even != nil && even.Next != nil {
		odd.Next = odd.Next.Next
		even.Next = even.Next.Next
		odd = odd.Next
		even = even.Next
	}
	odd.Next = evenHead
	return head
}

type ListNode struct {
	Val  int
	Next *ListNode
}
